image_db_to_file/requirements.txt
mysql.connector-python
logging
